// app/frontend/src/types.ts
export interface CustomStyles {
    AccentHigh: string;
    AccentLight: string;
    AccentDark: string;
    TextColor: string;
    BackgroundColor: string;
    ForegroundColor: string;
    FormBackgroundColor: string;
    BorderRadius: string;
    BorderWidth: string;
    FontBaseSize: string;
}

export interface CustomText {
    chatTitle: string;
    chatSubtitle: string;
    exampleQuestions: string[];
}